Make SDL work on Windows
