+++ 3rdparty/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(imgui)
 
-if (USE_SDL)
+if (USE_SDL OR USE_SDL_MIXER)
     add_subdirectory(SDL)
 endif()
