+++ src/GUI/WindowSystem.h
@@ -1,0 +1,92 @@
+#pragma once
+
+#include <string>
+#include <config.h>
+
+#if defined(USE_SDL) || defined(USE_SDL_MIXER)
+    #include <SDL3/SDL.h>
+#endif
+
+#ifdef USE_SDL
+    #include <SDL3/SDL_vulkan.h>
+#else
+    #include <windows.h>
+    #define VK_USE_PLATFORM_WIN32_KHR
+#endif
+
+#include "Vulkan/VkAdapter.h"
+
+class WindowSystem
+{
+public:
+#ifdef USE_SDL
+    WindowSystem();
+#else
+    WindowSystem(HINSTANCE hInstance, int nShowCmd);
+#endif // USE_SDL
+
+    virtual ~WindowSystem();
+
+    void createWindow(VkAdapter* vkAdapter);
+    void destroyWindow();
+
+    void render();
+
+    void minimizeWindow();
+    void maximizeRestoreWindow();
+    void closeWindow();
+
+    bool quit() const { return _quit; }
+
+    float getMainScale() const { return _mainScale; }
+    bool borderlessWindow() const { return _borderlessWindow; }
+
+    float titleBarHeight() const { return _titlebarHeight; }
+    float windowButtonWidth() const { return _buttonWidth; }
+
+    const char* windowTitle() const;
+
+    void getVkInstanceExtensions(std::vector<const char*>& extensions) const;
+
+    void createVkSurfaceKHR(
+        VkInstance instance,
+        VkSurfaceKHR* surface, int *width, int *height) const;
+
+private:
+    VkAdapter* _vkAdapter = nullptr;
+
+    float _mainScale = 1.f;
+    bool _borderlessWindow = true;
+    // TODO: make scale aware
+    float _titlebarHeight = 32.f;
+    float _buttonWidth = 32.f;
+    float _totalButtonWidth = 3 * 32.f;
+
+    bool _quit = false;
+
+#ifdef USE_SDL
+    void sdlWndProc(SDL_Event& event);
+    static SDL_HitTestResult SDLCALL sdlHitTest(SDL_Window* win, const SDL_Point* area, void* data);
+    static void SDLCALL sdlCallbackOpenFile(void* userdata, const char* const* filelist, int filter);
+
+    SDL_Window* _sdlWindow;
+    bool _isMaximized = false;
+#else
+    // WIN32 stuff for working with borderless windows
+    // see https://github.com/melak47/BorderlessWindow/tree/main
+    static LRESULT CALLBACK w32WndProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam);
+    bool w32CompositionEnabled();
+    DWORD w32Style();
+    void w32CreateWindow(int nShowCmd);
+    void w32SetBorderless(bool borderless);
+    bool w32IsMaximized();
+    void w32AdjustMaximizedClientRect(RECT& rect);
+    LRESULT w32HitTest(POINT cursor) const;
+    std::string w32OpenFileName(const char* title, const char* initialDir, const char* filter, bool multiSelect);
+
+    HWND _hwnd;
+    HINSTANCE _hInstance;
+    int _nShowCmd;
+#endif
+};
+
